Fix/lighten history browsing
This fixes #2921 , essentially by replacing a call to `/graph` (that is now restricted) to `/rev/{id}/log`, that has the "nice" side-effect to pre-fill the cache because those revisions' metadata is fetched by other artifacts afterwards.

We change gears compared to the previous version: this only loads \~1000 history entries, in a single API call. otherwise it's too complicated, too slow... loading the whole history is not worth the difficulty.

by the way we remove the "loading status" file, and parallel loading entirely, because it's too complicated and too long anyway; instead we remove loading the entire history when using the Web API.